requirements.txt
attr==0.3.2
attrs==25.2.0
coacd==1.0.5
colorama==0.4.6
Flask==3.1.0
grpcio==1.64.0
httpx==0.28.1
jsonschema==4.23.0
matplotlib==3.8.3
numpy==1.26.4
openai==1.66.2
pandas==2.2.1
paramiko==3.4.1
Pillow==11.1.0
protobuf==3.20.3
pygltflib==1.16.2
Pyqtree==1.0.0
Requests==2.32.3
scipy==1.15.2
Shapely==2.0.7
scikit-image==0.24.0
spacy==3.8.4
sshtunnel==0.4.0
trimesh==4.4.7
